# libs/clist: remove all sync.WaitGroup fields