There're corresponding channels, which can be used to do the same job
of all sync.WaitGroup fields. So, it's not necessary to use them
anymore.

Benchmark result for runtime speed can be noisy, but the improvement in
memory is clearly visible:

```
name         old time/op    new time/op    delta
Detaching-8    48.9ns ±13%    57.8ns ±17%     ~     (p=0.095 n=5+5)
Removed-8       118ns ±39%      59ns ±27%  -50.19%  (p=0.008 n=5+5)
PushBack-8      224ns ±22%     167ns ±22%  -25.17%  (p=0.032 n=5+5)

name         old alloc/op   new alloc/op   delta
Detaching-8     0.00B          0.00B          ~     (all equal)
Removed-8       0.00B          0.00B          ~     (all equal)
PushBack-8       327B ± 0%      279B ± 0%  -14.68%  (p=0.000 n=5+4)

name         old allocs/op  new allocs/op  delta
Detaching-8      0.00           0.00          ~     (all equal)
Removed-8        0.00           0.00          ~     (all equal)
PushBack-8       5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.008 n=5+5)
```

Fixes #6450