@tychoish 2021-05-11
I just ran the mempool's benchmarks which are one level higher in the stack, and may provide a better idea of what the impact of this change is in context of the code that actually uses the linked list implementation: 

I used the following command: 

```
go test -cpu=8 -benchmem -bench=. -run="Bench.*" ./mempool
```

(Changing the number of CPUs didn't change things too much)

First on master as of today: 

```
go test -cpu=8 -benchmem -bench=. -run="Bench.*" ./mempool
goos: linux
goarch: amd64
pkg: github.com/tendermint/tendermint/mempool
cpu: AMD Ryzen 9 3950X 16-Core Processor            
BenchmarkReap-8                     1274            892478 ns/op          485780 B/op      10001 allocs/op
BenchmarkCheckTx-8                362602              3051 ns/op            1521 B/op         32 allocs/op
BenchmarkParallelCheckTx-8        220724              5407 ns/op            1574 B/op         32 allocs/op
BenchmarkCheckDuplicateTx-8       317997              3862 ns/op            1547 B/op         33 allocs/op
BenchmarkCacheInsertTime-8       2636832               421.1 ns/op            80 B/op          2 allocs/op
BenchmarkCacheRemoveTime-8       3553003               324.1 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/tendermint/tendermint/mempool        12.244s
go test -cpu=8 -benchmem -bench=. -run="Bench.*" ./mempool  19.47s user 0.71s system 157% cpu 12.840 total
```

Then with this change: 

```
go test -cpu=8 -benchmem -bench=. -run="Bench.*" ./mempool
goos: linux
goarch: amd64
pkg: github.com/tendermint/tendermint/mempool
cpu: AMD Ryzen 9 3950X 16-Core Processor            
BenchmarkReap-8                     1180            903505 ns/op          485781 B/op      10001 allocs/op
BenchmarkCheckTx-8                342006              3048 ns/op            1477 B/op         30 allocs/op
BenchmarkParallelCheckTx-8        234578              5487 ns/op            1518 B/op         30 allocs/op
BenchmarkCheckDuplicateTx-8       308164              3725 ns/op            1503 B/op         31 allocs/op
BenchmarkCacheInsertTime-8       2632506               439.7 ns/op            80 B/op          2 allocs/op
BenchmarkCacheRemoveTime-8       3492236               337.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/tendermint/tendermint/mempool        12.471s
```

The impacts that I see: 
- this change omits a few mallocs likely because we're not allocating the wait group object any more. This is nice, but I wouldn't worry too much about a few mallocs. 
- the reaping benchmark, which is the slowest operation in the mempool, actually gets a bit slower (but without changing memory use.) Maybe about 1% I don't have a great explanation for this except maybe 'channels are slightly slower than waitgroups for raw performance?"
- CheckTX, ParalletCheckTX , and DuplicateCheckTX all all get about 1% faster, though these operations were always pretty fast, and 2 orders of magnittude faster than reaping to start with. Most of the speed of the CheckTx operation depends on the application, and the test application is very fast. The speed-up here is less likely to be noticed because the application's performance is the larger factor here in practice.