@tychoish 2021-05-11
Based on my hunch that "for pure performance, wait groups are likely a touch faster than channels," I tried the "opposite" approach, where I removed the channels, (see [this patch](https://gist.github.com/805b16eb8f6ea906b8f7779b0f2f97b2)) and left the wait group as the primary concurrency control.

Just for good measure, the current state of master: 

```
goos: linux
goarch: amd64
pkg: github.com/tendermint/tendermint/mempool
cpu: AMD Ryzen 9 3950X 16-Core Processor            
BenchmarkReap-8                     1326            890573 ns/op          485779 B/op      10001 allocs/op
BenchmarkCheckTx-8                344700              2985 ns/op            1525 B/op         32 allocs/op
BenchmarkParallelCheckTx-8        221882              5659 ns/op            1573 B/op         32 allocs/op
BenchmarkCheckDuplicateTx-8       301230              3808 ns/op            1553 B/op         33 allocs/op
BenchmarkCacheInsertTime-8       2787404               403.6 ns/op            81 B/op          2 allocs/op
BenchmarkCacheRemoveTime-8       3488594               327.5 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/tendermint/tendermint/mempool        12.266s
```


No stored channels, just the waitgroup: 

```
goos: linux
goarch: amd64
pkg: github.com/tendermint/tendermint/mempool
cpu: AMD Ryzen 9 3950X 16-Core Processor            
BenchmarkReap-8                     1191            892324 ns/op          485781 B/op      10001 allocs/op
BenchmarkCheckTx-8                398050              3170 ns/op            1307 B/op         30 allocs/op
BenchmarkParallelCheckTx-8        235194              5431 ns/op            1357 B/op         30 allocs/op
BenchmarkCheckDuplicateTx-8       285861              4000 ns/op            1351 B/op         31 allocs/op
BenchmarkCacheInsertTime-8       2613043               424.2 ns/op            80 B/op          2 allocs/op
BenchmarkCacheRemoveTime-8       3539300               343.7 ns/op             0 B/op          0 allocs/op
PASS
ok      github.com/tendermint/tendermint/mempool        11.670s
```

Again the results are mixed, but the performance on reaping doesn't have a particularly big degradation, so that's nice. It's a touch faster for the parallel checkTx (maybe 2%) but a touch slower for the other checktx (with fewer allocs). I think the parallel benchmark is perhaps the most important of the checktx operations. It's also the case that the entire suite finishes faster and the throughput of the operations goes up, so I'm heartened by this a bit. 

My change was also very quick, and I suspect that my hack to keep the API the same (e.g. where I start a go routine to close the signal channel that I create,) is slower than it should be, and if we can think of different ways to handle this API then we might be able to get more speed, again, if we want it. 

------ 

Benchmarks are only a part of this story though: having two different concurrency control mechanisms is definitely more confusing than it needs to be, so I think it does improve the code a bit to change this.