@cuonglm 2021-05-12
@tychoish running benchmark just 1 time makes no much sense, though. Can you retry with `-count=5` or `-count=10` then doing a `benchstat` on them?

Example, in the PR description, you can see a boost in running time if you only look at the percent comparison:

```
name         old time/op    new time/op    delta
Detaching-8    48.9ns ±13%    57.8ns ±17%     ~     (p=0.095 n=5+5)
Removed-8       118ns ±39%      59ns ±27%  -50.19%  (p=0.008 n=5+5)
PushBack-8      224ns ±22%     167ns ±22%  -25.17%  (p=0.032 n=5+5)
```

But they have too much noise (e.g: 39% and 27% for Removed-8), thay's why I don't state that the performance have a boost.